Moved Block namespace
