+++ src/Sulu/Bundle/CoreBundle/Command/UpgradeInternalLinksCommand.php
@@ -13,8 +13,8 @@ namespace Sulu\Bundle\CoreBundle\Command;
 use PHPCR\PropertyType;
 use PHPCR\SessionInterface;
 use Sulu\Bundle\SnippetBundle\Snippet\SnippetRepository;
-use Sulu\Component\Content\Block\BlockPropertyInterface;
-use Sulu\Component\Content\Block\BlockPropertyWrapper;
+use Sulu\Component\Content\Types\Block\BlockPropertyInterface;
+use Sulu\Component\Content\Types\Block\BlockPropertyWrapper;
 use Sulu\Component\Content\Exception\ResourceLocatorNotFoundException;
 use Sulu\Component\Content\Mapper\ContentMapperInterface;
 use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
