+++ src/Sulu/Bundle/SearchBundle/Tests/Fixtures/SecondStructureCache.php
@@ -11,8 +11,8 @@
 namespace Sulu\Bundle\SearchBundle\Tests\Fixtures;
 
 use Sulu\Component\Content\Property;
-use Sulu\Component\Content\Block\BlockProperty;
-use Sulu\Component\Content\Block\BlockPropertyType;
+use Sulu\Component\Content\Types\Block\BlockProperty;
+use Sulu\Component\Content\Types\Block\BlockPropertyType;
 use Sulu\Component\Content\PropertyTag;
 use Sulu\Component\Content\Section\SectionProperty;
 use Sulu\Component\Content\Structure\Page;
