+++ src/Sulu/Component/Content/Template/Resources/Skeleton/StructureClass.php.twig
@@ -1,8 +1,8 @@
 <?php
 use Sulu\Component\Content\Property;
 use Sulu\Component\Content\PropertyParameter;
-use Sulu\Component\Content\Block\BlockProperty;
-use Sulu\Component\Content\Block\BlockPropertyType;
+use Sulu\Component\Content\Types\Block\BlockProperty;
+use Sulu\Component\Content\Types\Block\BlockPropertyType;
 use Sulu\Component\Content\PropertyTag;
 use Sulu\Component\Content\StructureTag;
 use Sulu\Component\Content\Section\SectionProperty;
