+++ tests/Sulu/Component/Content/Mapper/ContentMapperTest.php
@@ -15,8 +15,8 @@ use PHPCR\NodeInterface;
 use PHPCR\PropertyInterface;
 use ReflectionMethod;
 use Sulu\Bundle\TestBundle\Testing\PhpcrTestCase;
-use Sulu\Component\Content\Block\BlockProperty;
-use Sulu\Component\Content\Block\BlockPropertyType;
+use Sulu\Component\Content\Types\Block\BlockProperty;
+use Sulu\Component\Content\Types\Block\BlockPropertyType;
 use Sulu\Component\Content\BreadcrumbItemInterface;
 use Sulu\Component\Content\ContentEvents;
 use Sulu\Component\Content\Property;
